Added api call to get playlist
